Remove Model for Registrant from controllers
