+++ lib/controllers/api/v1/registration.js
@@ -3,6 +3,7 @@ const multer = require('multer');
 const mongoose = require('mongoose');
 const routes = require('express').Router();
 const Email = require('revolutionuc-emails');
+var Registrant = require('./../../../models/registrant');
 
 const email = new Email(process.env.MAILGUN_API_KEY, process.env.MAILGUN_DOMAIN);
 const MONGO_URL = process.env.MONGO_URL;
@@ -24,87 +25,6 @@ const storage = multer.diskStorage({
 const upload = multer({storage, limits: { fileSize: 20000000, files: 1 }}); // 20mb max
 const userCryptoAlgorithm = 'aes-256-ctr';
 const userCryptoPassword = process.env.MONGO_SALT;
-
-const registrantScheme = mongoose.Schema({
-  firstName: {
-    type: String,
-    required: true
-  },
-  lastName: {
-    type: String,
-    required: true
-  },
-  email: {
-    type: String,
-    required: true
-  },
-  school: {
-    type: String,
-    required: true
-  },
-  major: {
-    type: String,
-    required: true
-  },
-  education: {
-    type: String,
-    required: true
-  },
-  gender: {
-    type: String,
-    required: true
-  },
-  ethnicity: {
-    type: String,
-    required: true
-  },
-  howYouHeard: {
-    type: String
-  },
-  hackathons: {
-    type: String
-  },
-  shirtSize: {
-    type: String,
-    required: true
-  },
-  githubUsername: {
-    type: String
-  },
-  dob: {
-    type: String,
-    required: true
-  },
-  registrationTime: {
-    type: String,
-    required: true
-  },
-  checkedIn: {
-    type: Boolean,
-    default: false
-  },
-  emailVerified: {
-    type: Boolean,
-    default: false
-  },
-  attendanceConfirmed: {
-    type: Boolean,
-    default: false
-  },
-  dietaryRestrictions: {
-    type: String,
-    default: ''
-  },
-  accommodations: {
-    type: String,
-    default: ''
-  },
-  resumePath: {
-    type: String
-  }
-});
-
-const Registrant = mongoose.model('Registrant', registrantScheme);
 
 /**
  * Registration route handlers
